+++ tde/module/prompt_completions/documentation.lua
@@ -22,45 +22,38 @@
 --OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 --SOFTWARE.
 ]]
--- menu takes a bit of time to load in.
--- because of this we put it in the back so the rest of the system can already behave
--- Look into awesome-freedesktop for more information
+local icons = require("theme.icons")
+local docs = require('module.docs')
 
-require("module.menu")
+local basename = require("lib-tde.file").basename
 
-if not (general["disable_desktop"] == "1") then
-    if IsreleaseMode then
-        require("module.installer")
+
+local function create_response(doc_list)
+    local res = {}
+
+    for _, v in ipairs(doc_list) do
+        table.insert(res, {
+            icon = icons.docs,
+            text = string.gsub(basename(v), ".html", ""),
+            payload = v
+        })
     end
 
-    require("module.desktop")
+    return res
 end
 
-if IsreleaseMode then
-    require("tutorial")
-
-    require("module.dev-widget-update")
-
-    require("module.lua-completion")
-
-    require("lib-tde.signals").connect_exit(
-        function()
-            -- stop current autorun.sh
-            -- This is done because otherwise multiple instances would be running at the same time
-            awful.spawn("pgrep -f /etc/xdg/tde/autorun.sh | xargs kill -9")
-        end
-    )
-
-    local lockscreentime = general["screen_on_time"] or "120"
-    if general["screen_timeout"] == 1 or general["screen_timeout"] == nil then
-        awful.spawn("/etc/xdg/tde/autorun.sh " .. lockscreentime)
-    else
-        awful.spawn("/etc/xdg/tde/autorun.sh")
-    end
+local function get_completions(query)
+    return create_response(_G.docs(query, true))
 end
 
-require("module.docs")
-require("module.prompt")
+local function perform_action(payload)
+    docs.open_doc(payload)
+end
 
+local name = "Docs"
 
-require("module.screen_changed")
+return {
+    get_completion = get_completions,
+    perform_action = perform_action,
+    name = name,
+}